it's all context
