+++ test/ignite_command_runner_test.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
-import 'package:ignite_cli/commands/ignite_command.dart';
 import 'package:ignite_cli/flame_version_manager.dart';
 import 'package:ignite_cli/ignite_command_runner.dart';
+import 'package:ignite_cli/ignite_context.dart';
 import 'package:ignite_cli/version.g.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:mocktail/mocktail.dart';
